app1/requirements.txt
asgiref==3.8.1
async-timeout==5.0.1
beautifulsoup4==4.13.3
Django==5.1.6
django-admin-sortable2==2.2.4
django-appconf==1.1.0
django-compressor==4.5.1
django-htmlmin==0.11.0
django-js-asset==3.0.1
django-mptt==0.16.0
django-redis==5.4.0
html5lib==1.1
mysqlclient==2.2.7
pillow==11.1.0
python-dotenv==1.0.1
rcssmin==1.1.2
rjsmin==1.2.2
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
webencodings==0.5.1
